Refactor Move Ordering
